cambios paths en test+try/except
