+++ src/modeling.py
@@ -1,9 +1,10 @@
 import os
 import pickle
+import warnings
 
-def export_import_model(model:object, path_model:str, name:str, save:bool=True, open:bool=False):
+def export_import_model(model, path_model, name, save=True, open=False):
     '''
-    Funcion para exportar o importar el modelo entrenado.
+    Funcion para exportar o importar el modelo entrenado
 
     Parametros
     ----------
@@ -21,13 +22,22 @@ def export_import_model(model:object, path_model:str, name:str, save:bool=True, 
     # Exportamos el modelo con el nombre seleccionado, al path escogido.
     filename = os.path.join(path_model, name)
 
-    if save == True:
-        with open(filename, 'wb') as archivo_salida:
-            pickle.dump(model, archivo_salida)
+    if save:
+        try:
+            with open(filename, 'wb') as archivo_salida:
+                pickle.dump(model, archivo_salida)
+        except Exception as e:
+            print(f"Error al guardar el modelo: {str(e)}")
 
-    if open == True:
-        with open(filename, 'rb') as archivo_entrada:
-         model_pretrained = pickle.load(archivo_entrada)
+    if open:
+        try:
+            with open(filename, 'rb') as archivo_entrada:
+                model_pretrained = pickle.load(archivo_entrada)
+        except Exception as e:
+            print(f"Error al cargar el modelo: {str(e)}")
+            model_pretrained = None
+
+    return model_pretrained
 
 
 import warnings
